feat(Errors): Added ErrorBuilder typeclass but not yet used
