+++ src/Text/Gigaparsec/Errors/Token.hs
@@ -1,0 +1,4 @@
+{-# LANGUAGE Safe #-}
+module Text.Gigaparsec.Errors.Token (module Text.Gigaparsec.Errors.Token) where
+
+data Token = Raw !String | Named !String {-# UNPACK #-} !Word
